backend/requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
Django==4.2.4
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 
gunicorn==20.1.0 
djoser==2.1.0
drf_extra_fields==3.7.0.
Pillow==10.0.0
django-filter==23.2